+++ scvi/autotune/_tune.py
@@ -3,110 +3,11 @@ from __future__ import annotations
 import logging
 from typing import Any, Literal
 
-from ray.tune import Tuner
-
 from scvi._types import AnnOrMuData
 from scvi.autotune._experiment import AutotuneExperiment
 from scvi.model.base import BaseModelClass
 
 logger = logging.getLogger(__name__)
-
-
-def _trainable(
-    search_space: dict[str, callable],
-    experiment: AutotuneExperiment,
-) -> None:
-    """Implements a Ray Tune trainable function for an :class:`~scvi.autotune.AutotuneExperiment`.
-
-    Parameters
-    ----------
-    search_space
-        Hyperparameter configuration to evaluate. Note: this is different from
-        :attr:`~scvi.autotune.AutotuneExperiment.search_space` in that it is a single configuration
-        sampled from the search space, not the specification of the search space itself.
-    experiment
-        :class:`~scvi.autotune.AutotuneExperiment` to evaluate.
-
-    Notes
-    -----
-    See the Ray Tune
-    `documentation <https://docs.ray.io/en/latest/tune/api/trainable.html#function-trainable-api>`_
-    for more details.
-    """
-    from os.path import join
-
-    from lightning.pytorch.callbacks import Callback
-    from lightning.pytorch.loggers import TensorBoardLogger
-    from ray.train import get_context
-    from ray.tune.integration.pytorch_lightning import TuneReportCheckpointCallback
-
-    from scvi import settings
-
-    # This is to get around lightning import changes
-    tune_callback_cls = type(
-        "_TuneReportCheckpointCallback",
-        (TuneReportCheckpointCallback, Callback),
-        {},
-    )
-
-    callbacks = [
-        tune_callback_cls(metrics=experiment.metrics, on="validation_end", save_checkpoints=False)
-    ]
-    log_dir = join(
-        experiment.logging_dir,
-        experiment.name,
-        f"{get_context().get_trial_name()}_tensorboard",
-    )
-
-    model_args = search_space.get("model_args", {})
-    train_args = search_space.get("train_args", {})
-    train_args = {
-        "max_epochs": experiment.max_epochs,
-        "accelerator": "auto",
-        "devices": "auto",
-        "check_val_every_n_epoch": 1,
-        "enable_progress_bar": False,
-        "logger": TensorBoardLogger(log_dir),
-        "callbacks": callbacks,
-        **train_args,
-    }
-
-    settings.seed = experiment.seed
-    getattr(experiment.model_cls, experiment.setup_method_name)(
-        experiment.adata,
-        **experiment.setup_method_args,
-    )
-    model = experiment.model_cls(experiment.adata, **model_args)
-    model.train(**train_args)
-
-
-def _configure_tuner(experiment: AutotuneExperiment) -> Tuner:
-    """Configure a :class:`~ray.tune.Tuner` for an :class:`~scvi.autotune.AutotuneExperiment`."""
-    from ray.train import RunConfig
-    from ray.tune import with_parameters, with_resources
-    from ray.tune.tune_config import TuneConfig
-
-    trainable = with_parameters(_trainable, experiment=experiment)
-    trainable = with_resources(trainable, resources=experiment.resources)
-
-    tune_config = TuneConfig(
-        scheduler=experiment.scheduler,
-        search_alg=experiment.searcher,
-        num_samples=experiment.num_samples,
-    )
-    run_config = RunConfig(
-        name=experiment.name,
-        storage_path=experiment.logging_dir,
-        local_dir=experiment.logging_dir,
-        log_to_file=True,
-        verbose=1,
-    )
-    return Tuner(
-        trainable=trainable,
-        param_space=experiment.search_space,
-        tune_config=tune_config,
-        run_config=run_config,
-    )
 
 
 def run_autotune(
@@ -182,7 +83,8 @@ def run_autotune(
         If not provided, defaults to using all available resources. Note that fractional
         allocations are supported.
     experiment_name
-        Name of the experiment, used for logging purposes. Defaults to a unique ID.
+        Name of the experiment, used for logging purposes. Defaults to a unique ID concatenated
+        to the model class name.
     logging_dir
         Base directory to store experiment logs. Defaults to :attr:``scvi.settings.logging_dir``.
     scheduler_kwargs
@@ -224,5 +126,5 @@ def run_autotune(
     )
     logging.info(f"Running autotune experiment {experiment.name}.")
     init(log_to_driver=False)
-    experiment.result_grid = _configure_tuner(experiment).fit()
+    experiment.result_grid = experiment.get_tuner().fit()
     return experiment
